requirements.txt
Flask==3.0.0
PyMySQL==1.1.0
cryptography==41.0.7
Jinja2==3.1.2
Werkzeug==3.0.0
python-dotenv==1.0.0
